Remove mon_exploder from monster group in Machine mod
mon_exploder was a turret which was removed from the game in Jan 2018, since it hadn't been called to spawn anywhere in years. Since it is no longer defined, having it in the monster group throws an error.
Alternative fix- redefine mon_exploder based on the original turret and update it.