Update 0_spells.json
Rebalanced Psychiclysm's fear spell. It used to apply the fear for 3 hours and duration stacked, leading to it being pretty easy to wind up with the fear effect for days on end, which is just obnoxious. Also reduced the range slightly so those Mi-Go bastards can't snipe the player with it as easily.